fix nuxt tests
